Mcs5

+++ ICSharpCode.Decompiler/IL/ControlFlow/YieldReturnDecompiler.cs
@@ -45,6 +45,7 @@ namespace ICSharpCode.Decompiler.IL.ControlFlow
 		// for a description of this step.
 
 		ILTransformContext context;
+		ModuleDef metadata;
 
 		/// <summary>The type that contains the function being decompiled.</summary>
 		TypeDef currentType;
@@ -113,6 +114,7 @@ namespace ICSharpCode.Decompiler.IL.ControlFlow
 				return; // abort if enumerator decompilation is disabled
 			this.context = context;
 			this.currentType = function.CecilMethod.DeclaringType;
+			this.metadata = function.CecilMethod.Module;
 			this.enumeratorType = null;
 			this.enumeratorCtor = null;
 			this.stateField = null;
@@ -166,15 +168,18 @@ namespace ICSharpCode.Decompiler.IL.ControlFlow
 			newBody.SortBlocks(deleteUnreachableBlocks: true);
 			function.CheckInvariant(ILPhase.Normal);
 
-			if (!isCompiledWithMono) {
+			if (isCompiledWithMono)
+			{
+				CleanSkipFinallyBodies(function);
+			}
+			else
+			{
 				DecompileFinallyBlocks();
 				ReconstructTryFinallyBlocks(function);
 			}
 
 			context.Step("Translate fields to local accesses", function);
 			TranslateFieldsToLocalAccess(function, function, fieldToParameterMap, isCompiledWithMono);
-
-			CleanSkipFinallyBodies(function);
 
 			// On mono, we still need to remove traces of the state variable(s):
 			if (isCompiledWithMono) {
@@ -1069,26 +1074,62 @@ namespace ICSharpCode.Decompiler.IL.ControlFlow
 			if (skipFinallyBodies == null) {
 				return; // only mono-compiled code uses skipFinallyBodies
 			}
-			context.StepStartGroup("CleanSkipFinallyBodies", function);
-			Block entryPoint = AsyncAwaitDecompiler.GetBodyEntryPoint(function.Body as BlockContainer);
-			if (skipFinallyBodies.StoreInstructions.Count != 0 || skipFinallyBodies.AddressCount != 0) {
+			context.StepStartGroup("CleanFinallyBlocks", function);
+			if (skipFinallyBodies.StoreInstructions.Count != 0 || skipFinallyBodies.AddressCount != 0)
+			{
 				// misdetected another variable as doFinallyBodies?
 				// Fortunately removing the initial store of 0 is harmless, as we
 				// default-initialize the variable on uninit uses
 				return;
 			}
-			foreach (var tryFinally in function.Descendants.OfType<TryFinally>()) {
-				entryPoint = AsyncAwaitDecompiler.GetBodyEntryPoint(tryFinally.FinallyBlock as BlockContainer);
-				if (entryPoint?.Instructions[0] is IfInstruction ifInst) {
-					if (ifInst.Condition.MatchLogicNot(out var logicNotArg) && logicNotArg.MatchLdLoc(skipFinallyBodies)) {
+			foreach (var tryFinally in function.Descendants.OfType<TryFinally>())
+			{
+				if (!(tryFinally.FinallyBlock is BlockContainer container))
+					continue;
+				Block entryPoint = AsyncAwaitDecompiler.GetBodyEntryPoint(container);
+				if (entryPoint?.Instructions[0] is IfInstruction ifInst)
+				{
+					if (ifInst.Condition.MatchLogicNot(out var logicNotArg) && logicNotArg.MatchLdLoc(skipFinallyBodies))
+					{
 						context.Step("Remove if (skipFinallyBodies) from try-finally", tryFinally);
 						// condition will always be true now that we're using 'yield' instructions
 						entryPoint.Instructions[0] = ifInst.TrueInst;
 						entryPoint.Instructions.RemoveRange(1, entryPoint.Instructions.Count - 1);
+						// find new entryPoint after the old one was modified
+						entryPoint = AsyncAwaitDecompiler.GetBodyEntryPoint(container);
 					}
+				}
+				if (entryPoint?.Instructions.Count == 2
+					&& IsCallToMonoFinallyMethod(entryPoint.Instructions[0] as Call, out var finallyMethod)
+					&& entryPoint.Instructions[1].MatchLeave((BlockContainer)tryFinally.FinallyBlock))
+				{
+					context.Step("Inline " + finallyMethod.FullName + " into finally", tryFinally);
+					var finallyFunction = CreateILAst((MethodDef)finallyMethod.MetadataToken, context);
+					tryFinally.FinallyBlock = finallyFunction.Body;
+					var variables = finallyFunction.Variables.ToArray();
+					finallyFunction.Variables.Clear();
+					function.Variables.AddRange(variables);
 				}
 			}
 			context.StepEndGroup(keepIfEmpty: true);
+
+			bool IsCallToMonoFinallyMethod(Call call, out IMethod finallyMethod)
+			{
+				finallyMethod = default;
+				if (call == null)
+					return false;
+				if (!(call.Arguments.Count == 1 && call.Arguments[0].MatchLdThis()))
+					return false;
+				if (!call.Method.Name.StartsWith("<>__Finally"))
+					return false;
+				ITypeDefinition declaringTypeDefinition = call.Method.DeclaringTypeDefinition;
+				if (declaringTypeDefinition.MetadataToken != this.enumeratorType)
+					return false;
+				if (declaringTypeDefinition.ParentModule.PEFile.Module != metadata)
+					return false;
+				finallyMethod = call.Method;
+				return call.Method.MetadataToken is not null;
+			}
 		}
 	}
 }
